Refactoring (flake8 and imports)
