+++ src/evidently/dashboard/widgets/cat_target_pred_feature_table_widget.py
@@ -9,10 +9,11 @@ import pandas as pd
 import plotly.express as px
 
 from evidently import ColumnMapping
-from evidently.analyzers.cat_target_drift_analyzer import CatTargetDriftAnalyzer, QualityMetricsOptions
+from evidently.analyzers.cat_target_drift_analyzer import CatTargetDriftAnalyzer
 from evidently.model.widget import BaseWidgetInfo, AdditionalGraphInfo
 from evidently.dashboard.widgets.widget import Widget
 from evidently.dashboard.widgets.utils import CutQuantileTransformer
+from evidently.options import QualityMetricsOptions
 
 
 class CatTargetPredFeatureTable(Widget):
