+++ src/evidently/dashboard/widgets/class_confusion_based_feature_distr_table_widget.py
@@ -8,10 +8,11 @@ import pandas as pd
 import plotly.express as px
 
 from evidently import ColumnMapping
-from evidently.analyzers.classification_performance_analyzer import ClassificationPerformanceAnalyzer, QualityMetricsOptions
+from evidently.analyzers.classification_performance_analyzer import ClassificationPerformanceAnalyzer
 from evidently.model.widget import BaseWidgetInfo, AdditionalGraphInfo
 from evidently.dashboard.widgets.widget import Widget
 from evidently.dashboard.widgets.utils import CutQuantileTransformer
+from evidently.options import QualityMetricsOptions
 
 
 class ClassConfusionBasedFeatureDistrTable(Widget):
