+++ src/evidently/dashboard/widgets/data_drift_table_widget.py
@@ -10,10 +10,11 @@ import numpy as np
 import plotly.graph_objs as go
 
 from evidently import ColumnMapping
-from evidently.analyzers.data_drift_analyzer import DataDriftAnalyzer, DataDriftOptions, QualityMetricsOptions
+from evidently.analyzers.data_drift_analyzer import DataDriftAnalyzer
 from evidently.model.widget import BaseWidgetInfo, AlertStats, AdditionalGraphInfo
 from evidently.dashboard.widgets.widget import Widget, GREY, RED
 from evidently.dashboard.widgets.utils import CutQuantileTransformer
+from evidently.options import DataDriftOptions, QualityMetricsOptions
 
 
 class DataDriftTableWidget(Widget):
