+++ src/evidently/dashboard/widgets/num_output_drift_widget.py
@@ -8,10 +8,11 @@ import pandas as pd
 import plotly.figure_factory as ff
 
 from evidently import ColumnMapping
-from evidently.analyzers.num_target_drift_analyzer import NumTargetDriftAnalyzer, QualityMetricsOptions
+from evidently.analyzers.num_target_drift_analyzer import NumTargetDriftAnalyzer
 from evidently.model.widget import BaseWidgetInfo, AlertStats
 from evidently.dashboard.widgets.widget import Widget, GREY, RED
 from evidently.dashboard.widgets.utils import CutQuantileTransformer
+from evidently.options import QualityMetricsOptions
 
 
 class NumOutputDriftWidget(Widget):
