+++ src/evidently/dashboard/widgets/num_output_values_widget.py
@@ -10,9 +10,10 @@ import numpy as np
 import plotly.graph_objs as go
 
 from evidently import ColumnMapping
-from evidently.analyzers.num_target_drift_analyzer import NumTargetDriftAnalyzer, QualityMetricsOptions
+from evidently.analyzers.num_target_drift_analyzer import NumTargetDriftAnalyzer
 from evidently.model.widget import BaseWidgetInfo
 from evidently.dashboard.widgets.widget import Widget, GREY, RED
+from evidently.options import QualityMetricsOptions
 
 
 class NumOutputValuesWidget(Widget):
